refactor(openomni): relocate policy middleware to policy/ dir
