+++ packages/openomni/src/ingress/index.ts
@@ -3,5 +3,5 @@ export { IngressSessionResolver } from "./session-resolver.js";
 export { IngressEventProjector } from "./event-projector.js";
 export { SessionBridge } from "./session-bridge.js";
 export { IngressHandlers } from "./handlers.js";
-export { IngressAuthorityMiddleware } from "./middleware/ingress-authority.js";
+export { IngressAuthorityMiddleware } from "../policy/ingress-authority.js";
 export type { CoordinatorLike } from "./coordinator-like.js";
