+++ packages/openomni/src/subagent/index.ts
@@ -1,6 +1,6 @@
 export { SubagentRuntime } from "./runtime.js";
-export { SubagentSpawnPolicyMiddleware } from "./middleware/subagent-spawn-policy.js";
-export { BackgroundLimitsMiddleware } from "./middleware/background-limits.js";
+export { SubagentSpawnPolicyMiddleware } from "../policy/subagent-spawn-policy.js";
+export { BackgroundLimitsMiddleware } from "../policy/background-limits.js";
 export { SubagentConsultation } from "./consultation.js";
 export { BackgroundManager } from "./background-manager.js";
 export { BackgroundStore } from "./background-store.js";
